Create accountscrubber.py
This file is the core file that accountscrubber consists of. It will generate a few other files for housekeeping purposes, and should display everything to command line in a nice neat setup. This is written for Linux, so if you're running this on Windows, the script output may look bad (and I will not fix this).

This script also is not perfect. There isn't really a way for Praw to cleanly sort by your oldest posts and just eat through everything from there.